split main into multiple files
