+++ src/search.rs
@@ -1,6 +1,9 @@
-use crate::common::{Timing, random_sparse_vector, random_vector_name, retry_with_clients};
+use crate::args::Args;
+use crate::common::{
+    Timing, random_dense_vector, random_filter, random_sparse_vector, random_vector_name,
+    retry_with_clients,
+};
 use crate::processor::Processor;
-use crate::{Args, random_dense_vector, random_filter};
 use indicatif::ProgressBar;
 use qdrant_client::Qdrant;
 use qdrant_client::qdrant::point_id::PointIdOptions;
